% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block_for_reclin.R
\name{blocking}
\alias{blocking}
\title{Main function for blocking records given text data}
\usage{
blocking(
  text,
  distance = "cosine",
  M = 25,
  ef = 200,
  verbose = T,
  progress = "bar",
  n_threads = 4
)
}
\arguments{
\item{text}{input text data}

\item{distance}{distance metric for \code{RcppHNSW}}

\item{M}{see \code{RcppHNSW}}

\item{ef}{see \code{RcppHNSW}}

\item{verbose}{see \code{RcppHNSW}}

\item{progress}{see \code{RcppHNSW}}

\item{n_threads}{see \code{RcppHNSW}}
}
\value{
Returns vector with cluster indicators
}
\description{
Function that creates shingles (strings with 2 characters), applies approximate nearest neibhour search using
\code{RcppHNSW} and creates blocks using \code{igraph}
}
\details{
Imports
}
\author{
Maciej Beręsewicz
}
