% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{census}
\alias{census}
\title{Fictional census data}
\format{
A \code{data.table} with 25343 records. Each row represents one record, with the following columns:
\itemize{
\item{\code{PERSON_ID} -- a unique number for each person, consisting of postcode, house number and person number,}
\item{\code{PERNAME1} -- forename,}
\item{\code{PERNAME2} -- surname,}
\item{\code{SEX} -- gender (M/F),}
\item{\code{DOB_DAY} -- day of birth,}
\item{\code{DOB_MON} -- month of birth,}
\item{\code{DOB_YEAR} -- year of birth,}
\item{\code{hse_num} -- house number, a numeric label for each house within a street,}
\item{\code{ENUMCAP} -- an address consisting of house number and street name,}
\item{\code{ENUMPC} -- postcode,}
\item{\code{str_nam} -- street name of person's household's street,}
\item{\code{cap_add} -- full address, consisting of house number, street name and postcode,}
\item{\code{CENSUS_ID} -- person ID with "CENS" added in front.}
}
}
\usage{
census
}
\description{
This data set was created by Paula McLeod, Dick Heasman and Ian Forbes, ONS,
for the ESSnet DI on-the-job training course, Southampton, 25-28 January 2011.
It contains fictional data representing some observations from a decennial Census.
}
\examples{

data("census")
head(census)

}
\references{
McLeod, P., Heasman, D., Forbes, I. (2011). Simulated data for the ESSnet DI on-the-job training course,
Southampton, 25-28 January 2011.
\url{https://wayback.archive-it.org/12090/20231221144450/https://cros-legacy.ec.europa.eu/content/job-training_en}
}
\keyword{datasets}
