% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controls.R
\name{control_annoy}
\alias{control_annoy}
\title{Controls for Annoy}
\usage{
control_annoy(n_trees = 250, build_on_disk = FALSE, ...)
}
\arguments{
\item{n_trees}{An integer specifying the number of trees to build in the Annoy index.}

\item{build_on_disk}{A logical value indicating whether to build the Annoy index on disk instead of in memory.}

\item{...}{Additional arguments.}
}
\value{
Returns a list with parameters.
}
\description{
Controls for Annoy algorithm used in the package (see \link{RcppAnnoy} for details).
}
