% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controls.R
\name{control_kd}
\alias{control_kd}
\title{Controls for KD}
\usage{
control_kd(
  algorithm = "dual_tree",
  epsilon = 0,
  leaf_size = 20,
  random_basis = FALSE,
  rho = 0.7,
  tau = 0,
  tree_type = "kd",
  ...
)
}
\description{
Controls for KD
}
