% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controls.R
\name{control_lsh}
\alias{control_lsh}
\title{Controls for LSH}
\usage{
control_lsh(
  bucket_size = 10,
  hash_width = 6,
  num_probes = 5,
  projections = 10,
  tables = 30,
  ...
)
}
\arguments{
\item{bucket_size}{The size of a bucket in the second level hash.}

\item{hash_width}{The hash width for the first-level hashing in the LSH preprocessing.}

\item{num_probes}{Number of additional probes for multiprobe LSH.}

\item{projections}{The number of hash functions for each table.}

\item{tables}{The number of hash tables to be used.}

\item{...}{Additional arguments.}
}
\value{
Returns a list with parameters.
}
\description{
Controls for LSH algorithm used in the package (see \link[mlpack]{lsh} for details).
}
