% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controls.R
\name{controls_ann}
\alias{controls_ann}
\title{Controls for approximate nearest neighbours algorithms}
\usage{
controls_ann(
  sparse = FALSE,
  k_search = 30,
  nnd = control_nnd(),
  hnsw = control_hnsw(),
  lsh = control_lsh(),
  kd = control_kd(),
  annoy = control_annoy()
)
}
\arguments{
\item{sparse}{whether sparse data should be used as an input for algorithms,}

\item{k_search}{number of neighbours to search,}

\item{nnd}{parameters for \link[rnndescent:rnnd_build]{rnnd_build()} and \link[rnndescent:rnnd_query]{rnnd_query()} (should be inside \code{\link[=control_nnd]{control_nnd()}} function),}

\item{hnsw}{parameters for \link[RcppHNSW:hnsw_build]{hnsw_build()} and \link[RcppHNSW:hnsw_search]{hnsw_search()} (should be inside \code{\link[=control_hnsw]{control_hnsw()}} function),}

\item{lsh}{parameters for \link[mlpack:lsh]{lsh()} function (should be inside \code{\link[=control_lsh]{control_lsh()}} function),}

\item{kd}{kd parameters for \link[mlpack:knn]{knn()} function (should be inside \code{\link[=control_kd]{control_kd()}} function),}

\item{annoy}{parameters for the \link[RcppAnnoy]{RcppAnnoy} package (should be inside \code{\link[=control_annoy]{control_annoy()}} function).}
}
\value{
Returns a list with parameters.
}
\description{
Controls for ANN algorithms used in the package.
}
\author{
Maciej Beręsewicz
}
