% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controls.R
\name{controls_lsh}
\alias{controls_lsh}
\title{Controls for LSH}
\usage{
controls_lsh(
  bucket_size = 500,
  hash_width = 10,
  num_probes = 0,
  projections = 10,
  tables = 30,
  ...
)
}
\description{
Controls for LSH
}
