% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controls.R
\name{controls_txt}
\alias{controls_txt}
\title{Controls for processing text data}
\usage{
controls_txt(
  n_shingles = 2L,
  n_chunks = 10L,
  lowercase = TRUE,
  strip_non_alphanum = TRUE
)
}
\arguments{
\item{n_shingles}{length of shingles (default \code{2L}),}

\item{n_chunks}{passed to (default \code{10L}),}

\item{lowercase}{should the characters be made lowercase? (default \code{TRUE}),}

\item{strip_non_alphanum}{should punctuation and white space be stripped? (default \code{TRUE}).}
}
\value{
Returns a list with parameters.
}
\description{
Controls for text data used in the \code{blocking} function (if \code{representation = shingles}), passed to \link[tokenizers:shingle-tokenizers]{tokenizers::tokenize_character_shingles}.
}
\author{
Maciej Beręsewicz
}
