% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_errors.R
\name{estimate_errors}
\alias{estimate_errors}
\title{Estimate record linkage errors}
\usage{
estimate_errors(
  x = NULL,
  y = NULL,
  blocking_result = NULL,
  n = NULL,
  N = NULL,
  G,
  alpha = NULL,
  p = NULL,
  lambda = NULL,
  tol = 10^(-10),
  maxiter = 1000
)
}
\description{
Estimate record linkage errors
}
