% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{foreigners}
\alias{foreigners}
\title{Fictional 2024 population of foreigners in Poland}
\format{
A \code{data.table} with 110000 records. Each row represents one record, with the following columns:
\itemize{
\item{\code{fname} -- first name,}
\item{\code{sname} -- second name,}
\item{\code{surname} -- surname,}
\item{\code{date} -- date of birth,}
\item{\code{region} -- region (county),}
\item{\code{country} -- country,}
\item{\code{true_id} -- person ID.}
}
}
\usage{
foreigners
}
\description{
A fictional data set of the foreign population in Poland,
generated based on publicly available information
while maintaining the distributions from administrative registers.
}
\examples{

data("foreigners")
head(foreigners)

}
\keyword{datasets}
