% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval.R
\name{get_confusion}
\alias{get_confusion}
\title{Confusion matrix}
\usage{
get_confusion(TP, FP, FN, TN)
}
\arguments{
\item{TP}{TP}

\item{FP}{FP}

\item{FN}{FN}

\item{TN}{TN}
}
\value{
Returns a confusion matrix.
}
\description{
Function creates a confusion matrix from raw counts.
}
