% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval.R
\name{get_metrics}
\alias{get_metrics}
\title{Metrics for evaluating dedupliaction and record linkage}
\usage{
get_metrics(TP, FP, FN, TN)
}
\arguments{
\item{TP}{TP}

\item{FP}{FP}

\item{FN}{FN}

\item{TN}{TN}
}
\value{
Returns a list containing evaluation metrics.
}
\description{
Function calculates standard evaluation metrics.
}
