% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_annoy.R
\name{method_annoy}
\alias{method_annoy}
\title{An internal function to use Annoy algorithm via the \link[RcppAnnoy]{RcppAnnoy} package.}
\usage{
method_annoy(x, y, k, distance, verbose, path, seed, control)
}
\arguments{
\item{x}{deduplication or reference data,}

\item{y}{query data,}

\item{k}{number of neighbours to return,}

\item{distance}{distance metric,}

\item{verbose}{if TRUE, log messages to the console,}

\item{path}{path to write the index,}

\item{seed}{seed for the pseudo-random numbers algorithm,}

\item{control}{controls for \code{new} or \code{build} methods for \link[RcppAnnoy]{RcppAnnoy}.}
}
\description{
See details of the \link[RcppAnnoy]{RcppAnnoy} package.
}
\author{
Maciej Beręsewicz
}
