% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_hnsw.R
\name{method_hnsw}
\alias{method_hnsw}
\title{An internal function to use HNSW algorithm via the RcppHNSW package.}
\usage{
method_hnsw(x, y, k, distance, verbose, n_threads, path, control, seed)
}
\arguments{
\item{x}{deduplication or reference data,}

\item{y}{query data,}

\item{k}{number of neighbours to return,}

\item{distance}{type of distance to calculate,}

\item{verbose}{if TRUE, log messages to the console,}

\item{n_threads}{Maximum number of threads to use,}

\item{path}{path to write the index,}

\item{control}{controls for the HNSW algorithm.}
}
\description{
See details of \link[RcppHNSW]{hnsw_build} and \link[RcppHNSW]{hnsw_search}.
}
\author{
Maciej Beręsewicz
}
\keyword{internal}
