% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_mlpack.R
\name{method_mlpack}
\alias{method_mlpack}
\title{An internal function to use algorthms from the mlpack package.}
\usage{
method_mlpack(x, y, algo = c("lsh", "kd"), k, verbose, seed, path, control)
}
\arguments{
\item{x}{Deduplication or reference data.}

\item{y}{Query data.}

\item{algo}{Which algorithm should be used. Possible: \code{lsh} or \code{kd}.}

\item{k}{Number of neighbors to return.}

\item{verbose}{If TRUE, log messages to the console.}

\item{seed}{seed for the pseudo-random numbers algorithm.}

\item{path}{path to write the index.}

\item{control}{controls for  \code{lsh} or \code{kd}.}
}
\description{
See details of \link[mlpack:lsh]{mlpack::lsh} and \link[mlpack:knn]{mlpack::knn}
}
\author{
Maciej Beręsewicz
}
