% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RLdata500}
\alias{RLdata500}
\title{\code{RLdata500} data from \pkg{RecordLinkage} package}
\format{
A \code{data.table} with 500 records. Each row represents one record, with the following columns:

\itemize{
\item{\code{fname_cq} -- first name, first component,}
\item{\code{fname_c2} -- first name, second component,}
\item{\code{lname_c1} -- last name, first component,}
\item{\code{lname_c2} -- last name, second component,}
\item{\code{by} -- year of birth,}
\item{\code{bm} -- month of birth,}
\item{\code{bd} -- day of birth,}
\item{\code{rec_id} -- record id,}
\item{\code{ent_id} -- entity id.}
}
}
\usage{
RLdata500
}
\description{
This data is taken from \pkg{RecordLinkage} R package developed by Murat Sariyar and Andreas Borg.
The package is licensed under GPL-3 license.

The \code{RLdata500} table contains artificial personal data.
Some records have been duplicated with randomly generated errors. \code{RLdata500} contains fifty duplicates.
}
\examples{

data("RLdata500")
head(RLdata500)

}
\references{
Sariyar M., Borg A. (2022). RecordLinkage: Record Linkage Functions for Linking and Deduplicating Data Sets.
R package version 0.4-12.4, \url{https://CRAN.R-project.org/package=RecordLinkage}
}
\keyword{datasets}
