% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentence_to_vector.R
\name{sentence_to_vector}
\alias{sentence_to_vector}
\title{Sentence to vector}
\usage{
sentence_to_vector(sentences, model)
}
\arguments{
\item{sentences}{a character vector,}

\item{model}{a matrix containing word embeddings (e.g., GloVe).}
}
\description{
Function creates a matrix with word embeddings using a given model.
}
