% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blocking.R
\name{blocking}
\alias{blocking}
\title{Main function for blocking records given text data}
\usage{
blocking(
  x,
  y = NULL,
  ann = c("hnsw", "lsh", "annoy", "kd"),
  distance = c("cosine", "euclidean", "l2", "ip", "manhatan", "hamming", "angular"),
  verbose = c(0, 1, 2),
  n_threads = 1,
  seed = 2023,
  control_txt = controls_txt(),
  control_ann = controls_ann()
)
}
\arguments{
\item{x}{input text or matrix data,}

\item{y}{input text or matrix data (default NULL),}

\item{ann}{algorithm to be used for searching for ann (possible, \code{c("hnsw", "lsh", "annoy", "kd")},
default \code{"hnsw"}),}

\item{distance}{distance metric (default \code{cosine}),}

\item{verbose}{whether log should be provided (0 = none, 1 = main, 2 = ann algorithms),}

\item{n_threads}{number of threads used for the ann,}

\item{seed}{seed for the algorithms,}

\item{control_txt}{list of controls for text data,}

\item{control_ann}{list of controls for ann algorithms.}
}
\value{
Returns list with cluster indicators for x, y, ann method and colnames of the dtm
}
\description{
Function that creates shingles (strings with 2 characters), applies approximate nearest neibhour search using
\code{RcppHNSW} and creates blocks using \code{igraph}
}
\details{
Imports
}
\examples{
df_example <- data.frame(txt = c("jankowalski", "kowalskijan", "kowalskimjan",
"kowaljan", "montypython", "pythonmonty", "cyrkmontypython", "monty"))

result <- blocking(x = df_example$txt,
                   ann = "hnsw",
                   control_ann = controls_ann(hnsw = list(M = 5, ef_c = 10, ef_s = 10)))


}
\author{
Maciej Beręsewicz
}
