% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blocking.R
\name{blocking}
\alias{blocking}
\title{Main function for blocking records given text data}
\usage{
blocking(
  x,
  y = NULL,
  deduplication = TRUE,
  on = NULL,
  on_blocking = NULL,
  ann = c("hnsw", "lsh", "annoy", "kd", "nnd"),
  distance = c("cosine", "euclidean", "l2", "ip", "manhatan", "hamming", "angular"),
  ann_write = NULL,
  ann_colnames = NULL,
  true_blocks = NULL,
  verbose = c(0, 1, 2),
  graph = FALSE,
  seed = 2023,
  n_threads = 1,
  control_txt = controls_txt(),
  control_ann = controls_ann()
)
}
\arguments{
\item{x}{reference data (character vector or a matrix),}

\item{y}{query data (types the same), if not provided NULL by default,}

\item{deduplication}{whether deduplication should be applied (default TRUE as y is set to NULL),}

\item{on}{variables for ann search (currently not supported),}

\item{on_blocking}{variables for blocking (currently not supported),}

\item{ann}{algorithm to be used for searching for ann (possible, \code{c("hnsw", "lsh", "annoy", "kd")}, default \code{"hnsw"}),}

\item{distance}{distance metric (default \code{cosine}),}

\item{ann_write}{writing an index to file. Two files will be created: 1) an index, 2) and txt file with column names (currently not supported),}

\item{ann_colnames}{testing}

\item{true_blocks}{matrix with true blocks to calculate evaluation metrics (all metrics from \code{\link[igraph:compare]{igraph::compare()}} are returned).}

\item{verbose}{whether log should be provided (0 = none, 1 = main, 2 = ann algorithms),}

\item{graph}{whether a graph should be returned,}

\item{seed}{seed for the algorithms,}

\item{n_threads}{number of threads used for the ann,}

\item{control_txt}{list of controls for text data,}

\item{control_ann}{list of controls for ann algorithms.}
}
\value{
Returns a list with containing:\cr
\itemize{
\item{\code{result} -- \code{data.frame} with indices (rows) of x, y and block}
\item{\code{ann} -- name of the ann algorithm used,}
\item{\code{metrics} -- metrics, if \code{true_blocks} is provided,}
\item{\code{colnames} -- variable names (colnames) used for search.}
}
}
\description{
Function that creates shingles (strings with 2 characters), applies approximate nearest neighbour search using
RcppHNSW, \link{RcppAnnoy} and \link{mlpack} and creates blocks using \link{igraph}.
}
\details{
Imports
}
\examples{

## general example
df_example <- data.frame(txt = c("jankowalski", "kowalskijan", "kowalskimjan",
"kowaljan", "montypython", "pythonmonty", "cyrkmontypython", "monty"))

result <- blocking(x = df_example$txt,
                   ann = "hnsw",
                   control_ann = controls_ann(hnsw = list(M = 5, ef_c = 10, ef_s = 10)))

result

## an example with true blocks

}
\author{
Maciej Beręsewicz
}
