% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blocking.R
\name{blocking}
\alias{blocking}
\title{Main function for blocking records given text data}
\usage{
blocking(
  x,
  y = NULL,
  deduplication = FALSE,
  verbose = T,
  progress = "bar",
  n_threads = 1,
  control_txt = controls_txt(),
  control_ann = controls_ann()
)
}
\arguments{
\item{x}{input text or matrix data,}

\item{y}{input text or matrix data (default NULL),}

\item{deduplication}{generate pairs from only \code{x},}

\item{verbose}{whether progress should be provided,}

\item{progress}{how the progress should be presented,}

\item{n_threads}{number of threads used for the ann,}

\item{control_txt}{list of controls for text data,}

\item{control_ann}{list of controls for ann algorithms.}
}
\value{
Returns vector with cluster indicators.
}
\description{
Function that creates shingles (strings with 2 characters), applies approximate nearest neibhour search using
\code{RcppHNSW} and creates blocks using \code{igraph}
}
\details{
Imports
}
\examples{
df_example <- data.frame(txt = c("jankowalski", "kowalskijan", "kowalskimjan",
"kowaljan", "montypython", "pythonmonty", "cyrkmontypython", "monty"))

blocking(x = df_example$txt,
         control_ann = controls_ann(hsnw_M = 5, hnsw_ef = 10))


}
\author{
Maciej Beręsewicz
}
