% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blocking.R
\name{blocking}
\alias{blocking}
\title{Main function for blocking records given text data}
\usage{
blocking(
  x,
  y = NULL,
  deduplication = TRUE,
  on = NULL,
  on_blocking = NULL,
  ann = c("nnd", "hnsw", "annoy", "lsh", "kd"),
  distance = c("cosine", "euclidean", "l2", "ip", "manhatan", "hamming", "angular"),
  ann_write = NULL,
  ann_colnames = NULL,
  true_blocks = NULL,
  verbose = c(0, 1, 2),
  graph = FALSE,
  seed = 2023,
  n_threads = 1,
  control_txt = controls_txt(),
  control_ann = controls_ann()
)
}
\arguments{
\item{x}{reference data (a character vector or a matrix),}

\item{y}{query data (a character vector or a matrix), if not provided NULL by default and thus deduplication is performed,}

\item{deduplication}{whether deduplication should be applied (default TRUE as y is set to NULL),}

\item{on}{variables for ANN search (currently not supported),}

\item{on_blocking}{variables for blocking records before ANN search (currently not supported),}

\item{ann}{algorithm to be used for searching for ann (possible, \code{c("nnd", "hnsw", "annoy", "lsh", "kd")}, default \code{"nnd"} which corresponds to nearest neighbour descent method),}

\item{distance}{distance metric (default \code{cosine}, more options are possible see details),}

\item{ann_write}{writing an index to file. Two files will be created: 1) an index, 2) and text file with column names,}

\item{ann_colnames}{file with column names if \code{x} or \code{y} are indices saved on the disk (currently not supported),}

\item{true_blocks}{matrix with true blocks to calculate evaluation metrics (standard metrics based on confusion matrix as well as all metrics from \code{\link[igraph:compare]{igraph::compare()}} are returned).}

\item{verbose}{whether log should be provided (0 = none, 1 = main, 2 = ANN algorithm verbose used),}

\item{graph}{whether a graph should be returned (default FALSE),}

\item{seed}{seed for the algorithms (for reproducibility),}

\item{n_threads}{number of threads used for the ANN algorithms and adding data for index and query,}

\item{control_txt}{list of controls for text data (passed only to \link[text2vec:itoken]{text2vec::itoken_parallel} or \link[text2vec:itoken]{text2vec::itoken}),}

\item{control_ann}{list of controls for the ANN algorithms.}
}
\value{
Returns a list with containing:\cr
\itemize{
\item{\code{result} -- \code{data.table} with indices (rows) of x, y, block and distance between points}
\item{\code{method} -- name of the ANN algorithm used,}
\item{\code{metrics} -- metrics for quality assessment, if \code{true_blocks} is provided,}
\item{\code{colnames} -- variable names (colnames) used for search,}
\item{\code{graph} -- \code{igraph} class object.}
}
}
\description{
Function creates shingles (strings with 2 characters, default), applies approximate nearest neighbour (ANN) algorithms via the \link{rnndescent}, RcppHNSW, \link{RcppAnnoy} and \link{mlpack} packages,
and creates blocks using graphs via \link{igraph}.
}
\details{
Imports
}
\examples{

## an example using RcppHNSW
df_example <- data.frame(txt = c("jankowalski", "kowalskijan", "kowalskimjan",
"kowaljan", "montypython", "pythonmonty", "cyrkmontypython", "monty"))

result <- blocking(x = df_example$txt,
                   ann = "hnsw",
                   control_ann = controls_ann(hnsw = list(M = 5, ef_c = 10, ef_s = 10)))

result

## an example using mlpack::lsh

result_lsh <- blocking(x = df_example$txt,
                       ann = "lsh")

result_lsh
}
\author{
Maciej Beręsewicz
}
