% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controls.R
\name{control_lsh}
\alias{control_lsh}
\title{Controls for LSH}
\usage{
control_lsh(
  bucket_size = 10,
  hash_width = 6,
  num_probes = 5,
  projections = 10,
  tables = 30,
  ...
)
}
\description{
Controls for LSH
}
