% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controls.R
\name{controls_ann}
\alias{controls_ann}
\title{Controls for approximate nearest neighbours algorithms}
\usage{
controls_ann(
  sparse = FALSE,
  k_search = 30,
  nnd = control_nnd(),
  hnsw = control_hnsw(),
  lsh = control_lsh(),
  kd = control_kd(),
  annoy = control_annoy()
)
}
\arguments{
\item{sparse}{whether sparse data should be used as an input for algorithms,}

\item{k_search}{number of neighbours to search,}

\item{nnd}{parameters for \code{\link[rnndescent:rnnd_build]{rnndescent::rnnd_build()}} and \code{\link[rnndescent:rnnd_query]{rnndescent::rnnd_query()}} (should be inside \link{control_nnd} function),}

\item{hnsw}{parameters for \code{\link[RcppHNSW:hnsw_build]{RcppHNSW::hnsw_build()}} and \code{\link[RcppHNSW:hnsw_search]{RcppHNSW::hnsw_search()}} (should be inside \link{control_hnsw} function),}

\item{lsh}{parameters for \code{\link[mlpack:lsh]{mlpack::lsh()}} function (should be inside \link{control_lsh} function),}

\item{kd}{kd parameters for \code{\link[mlpack:knn]{mlpack::knn()}} function (should be inside \link{control_kd} function),}

\item{annoy}{parameters for \link{RcppAnnoy} package (should be inside \link{control_annoy} function).}
}
\value{
Returns a list with parameters.
}
\description{
Controls for ANN algorithms used in the package.
}
\author{
Maciej Beręsewicz
}
