% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controls.R
\name{controls_ann}
\alias{controls_ann}
\title{Controls for approximate nearest neighbours algoritms}
\usage{
controls_ann(
  hnsw = list(M = 25, ef_c = 200, ef_s = 200, grain_size = 1, byrow = TRUE),
  lsh = list(bucket_size = 500, hash_width = 10, num_probes = 0, projections = 10, tables
    = 30),
  annoy = list(n_trees = 250),
  kd = list(algorithm = "dual_tree", epsilon = 0, leaf_size = 20, random_basis = FALSE,
    rho = 0.7, tau = 0, tree_type = "kd")
)
}
\arguments{
\item{hnsw}{parameters for \code{\link[RcppHNSW:hnsw_build]{RcppHNSW::hnsw_build()}} and \code{\link[RcppHNSW:hnsw_search]{RcppHNSW::hnsw_search()}}.}

\item{lsh}{parameters for \code{\link[mlpack:lsh]{mlpack::lsh()}}.}

\item{annoy}{parameters for \link{RcppAnnoy} package.}

\item{kd}{parameters for \code{\link[mlpack:knn]{mlpack::knn()}} function.}
}
\value{
Returns a list with parameters
}
\description{
Controls for ANN algorithms used in the package
}
\author{
Maciej Beręsewicz
}
