% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controls.R
\name{controls_ann}
\alias{controls_ann}
\title{Controls for approximate nearest neighbours algoritms}
\usage{
controls_ann(k = 2L, distance = "cosine", hsnw_M = 25, hnsw_ef = 200)
}
\arguments{
\item{k}{k nearest neighbours}

\item{distance}{distance metric for \code{RcppHNSW}}

\item{hsnw_M}{see \code{RcppHNSW}}

\item{hnsw_ef}{see \code{RcppHNSW}}
}
\value{
Returns a list with parameters
}
\description{
Controls for ANN algorithms used in the package
}
\author{
Maciej Beręsewicz
}
