% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controls.R
\name{controls_ann}
\alias{controls_ann}
\title{Controls for approximate nearest neighbours algorithms}
\usage{
controls_ann(
  sparse = FALSE,
  k_search = 30,
  nnd = control_nnd(),
  hnsw = control_hnsw(),
  lsh = control_lsh(),
  kd = control_kd(),
  annoy = control_annoy()
)
}
\arguments{
\item{sparse}{whether sparse data should be used as an input for algorithms,}

\item{k_search}{number of neighbours to search,}

\item{nnd}{list of parameters for \code{\link[rnndescent:rnnd_build]{rnndescent::rnnd_build()}} and \code{\link[rnndescent:rnnd_query]{rnndescent::rnnd_query()}},}

\item{hnsw}{list of parameters for \code{\link[RcppHNSW:hnsw_build]{RcppHNSW::hnsw_build()}} and \code{\link[RcppHNSW:hnsw_search]{RcppHNSW::hnsw_search()}},}

\item{lsh}{list of parameters for \code{\link[mlpack:lsh]{mlpack::lsh()}} function,}

\item{kd}{list of kd parameters for \code{\link[mlpack:knn]{mlpack::knn()}} function,}

\item{annoy}{list of parameters for \link{RcppAnnoy} package.}
}
\value{
Returns a list with parameters
}
\description{
Controls for ANN algorithms used in the package
}
\author{
Maciej Beręsewicz
}
