% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controls.R
\name{controls_txt}
\alias{controls_txt}
\title{Controls for processing text data}
\usage{
controls_txt(n_shingles = 2L, n_chunks = 10L)
}
\arguments{
\item{n_shingles}{length of shingles (default 2L), passed to \link[tokenizers:shingle-tokenizers]{tokenizers::tokenize_character_shingles},}

\item{n_chunks}{passed to (default 10L) \link[tokenizers:shingle-tokenizers]{tokenizers::tokenize_character_shingles}.}
}
\value{
Returns a list with parameters.
}
\description{
Controls for text data used in the \code{blocking} functions
}
\author{
Maciej Beręsewicz
}
