% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_annoy.R
\name{method_annoy}
\alias{method_annoy}
\title{An internal function to use Annoy algorithm via RcppAnnoy.}
\usage{
method_annoy(x, y, k, distance, verbose, seed, control)
}
\arguments{
\item{x}{Deduplication or reference data.}

\item{y}{Query data.}

\item{k}{Number of neighbors to return.}

\item{distance}{distance metric}

\item{verbose}{If TRUE, log messages to the console.}

\item{seed}{seed for the pseudo-random numbers algorithm.}

\item{control}{controls for  \code{lsh} or \code{kd}.}
}
\description{
See details of the \link{RcppAnnoy} package.
}
\details{
Imports
}
\author{
Maciej Beręsewicz
}
