% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_hnsw.R
\name{method_hnsw}
\alias{method_hnsw}
\title{Internal function of hnsw via RcppHNSW}
\usage{
method_hnsw(x, y = NULL, k, distance, M, ef_c, ef_s, verbose = 0, n_threads)
}
\arguments{
\item{x}{x}

\item{y}{y}

\item{k}{k}

\item{distance}{distance}

\item{M}{m}

\item{ef_c}{ef}

\item{ef_s}{ef}

\item{verbose}{ver}

\item{n_threads}{trh}
}
\description{
Tu bedzie opis
}
\author{
Maciej Beręsewicz
}
