% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_hnsw.R
\name{method_hnsw}
\alias{method_hnsw}
\title{An internal function to use hnsw algorithm via RcppHNSW.}
\usage{
method_hnsw(x, y, k, distance, verbose, n_threads, path, control)
}
\arguments{
\item{x}{Deduplication or reference data.}

\item{y}{Query data.}

\item{k}{Number of neighbors to return.}

\item{distance}{Type of distance to calculate.}

\item{verbose}{If TRUE, log messages to the console.}

\item{n_threads}{Maximum number of threads to use.}

\item{path}{path to write the index.}

\item{control}{Controls for the HNSW algorithm}
}
\description{
See details of \link[RcppHNSW:hnsw_build]{RcppHNSW::hnsw_build} and \link[RcppHNSW:hnsw_search]{RcppHNSW::hnsw_search}
}
\author{
Maciej Beręsewicz
}
