% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_nnd.R
\name{method_nnd}
\alias{method_nnd}
\title{An internal function to use the NN descent algorithm via the \link{rnndescent} package.}
\usage{
method_nnd(x, y, k, distance, verbose, n_threads, control)
}
\arguments{
\item{x}{deduplication or reference data,}

\item{y}{query data,}

\item{k}{number of neighbours to return,}

\item{distance}{type of distance to calculate,}

\item{verbose}{if TRUE, log messages to the console,}

\item{n_threads}{maximum number of threads to use,}

\item{control}{controls for the NN descent algorithm.}
}
\description{
See details of \link[rnndescent:rnnd_build]{rnndescent::rnnd_build} and \link[rnndescent:rnnd_query]{rnndescent::rnnd_query}.
}
\author{
Maciej Beręsewicz
}
