% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reclin2_pair_ann.R
\name{pair_ann}
\alias{pair_ann}
\title{Integration with the reclin2 package}
\usage{
pair_ann(
  x,
  y = NULL,
  on,
  deduplication = TRUE,
  keep_block = TRUE,
  add_xy = TRUE,
  ...
)
}
\arguments{
\item{x}{x}

\item{y}{y}

\item{on}{variable}

\item{deduplication}{deduplication}

\item{keep_block}{whether to keep block in the set}

\item{add_xy}{whether to add x and y}

\item{...}{arguments passed to \code{\link[=blocking]{blocking()}} function.}
}
\value{
Returns a \link{data.table} with two columns \code{.x} and \code{.y}. Columns \code{.x} and \code{.y} are row numbers from data.frames x and y respectively. This data.table is also of a class \code{pairs} which allows for integration witn the \code{\link[reclin2:compare_pairs]{reclin2::compare_pairs()}} package.
}
\description{
Function for the integration with the reclin2 package. The function is based on \code{\link[reclin2:pair_minsim]{reclin2::pair_minsim()}} and reuses some of its source code.
}
\details{
Imports
}
\examples{
data("linkexample1", "linkexample2", package = "reclin2")


}
\author{
Maciej Beręsewicz
}
