% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reclin2_pair_ann.R
\name{pair_ann}
\alias{pair_ann}
\title{Integration with reclin2 package}
\usage{
pair_ann(x, y, on, deduplication, ...)
}
\arguments{
\item{x}{x}

\item{y}{y}

\item{on}{variable}

\item{deduplication}{deduplication}

\item{...}{arguments passed to \code{\link[=blocking]{blocking()}} function.}
}
\value{
Returns a \link{data.table} with two columns \code{.x} and \code{.y}. Columns \code{.x} and \code{.y} are row numbers from data.frames x and y respectively. This data.table is also of a class \code{pairs} which allows for integration witn the \link{reclin2} package.
}
\description{
Integration with reclin2 package
}
\author{
Maciej Beręsewicz
}
